refactor: Add a manager for provider creation and registration
