+++ merino/configs/ci.toml
@@ -37,6 +37,7 @@ score = 0.3
 
 [ci.providers.wiki_fruit]
 # Whether or not this provider is enabled by default.
+type = "wiki_fruit"
 enabled_by_default = true
 
 [ci.providers.top_picks]
